Store served data and dump it on the exit.
We collect all data that is being served and consumed
by the test server and dump it on the exit when
server's destructor is called. This way we ensure
that data is printed in canonical way independent from
the order in which it was served/consumed.

Also, for queries that contain groupby we sort consumed
data so we do not depend on (unspecified) order in which
hashmap stores its entries.

As a bonus we get less noisy output when testing.

All tests has been updated to accomodate this change.