Clear up the confusion about stripes.
Made meaning of stripe more precise by
introducing a new term `query fragment`
that describes an abstract fragment of
a query that gets turned into stripe by
scheduler.

From now, stripe means one unit of work
that contains all the information needed
to perform the work like specific number
of file to be scanned (in case of stripe
from the first fragment).

The first fragment is an abstract fragment
of a query that has some information left
out like specific number of file in case
of the first fragment.

Fragments get replicated and turned into
stripes by the scheduler.